+++ src/FileOperationQueue.Core/Queue/FileOperationKind.cs
@@ -1,8 +1,0 @@
-namespace FileOperationQueue.Core.Queue;
-
-public enum FileOperationKind
-{
-    Copy = 0,
-    Move = 1
-}
-