+++ src/FileOperationQueue.Core/Worker/IFileOperationExecutor.cs
@@ -1,9 +1,0 @@
-using FileOperationQueue.Core.Queue;
-
-namespace FileOperationQueue.Core.Worker;
-
-public interface IFileOperationExecutor
-{
-    Task ExecuteAsync(FileOperationJob job, CancellationToken cancellationToken = default);
-}
-